+++ CHANGES.md
@@ -15,11 +15,12 @@ TUG improvements:
   * Check of digits after decimal separator for all coordinates, transmission ratios (gear + axle) and altitudes (Values in vehicle file changed to string values)
   * Vehicle file check included (For height and class code)
   * Expand vehicle file with vVehMax and GVMMax value
+  * Change vehicle file configuration parameter from "Rigid/Tractor" to "No/Yes"
   * Displacement of Genshape file into code
   * Add reference vehicle high in genshape class
   * Add minimum/maximum vehicle high in genshape class
   * Control of min/max height with vehicle height only in Declaration mode
-  * Calculation of new Result values (delta_CdxA_height; v_avg_LS/HS; t_amb_LS1)
+  * Calculation of new Result values (delta_CdxA_height; v_avg_LS/HS; t_amb_LS1, CdxA(ß)_H1/2, beta_H1/2)
   * Change of CdxA(ß) and beta calculation
   * Output files in Declaration mode extended
   * Output files adapted